Backend/app/core/deps.py
# protéger les routes
# user peut appeller les routes 
# le backend sait qui fait l appel

from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import jwt, JWTError
from app.core.config import settings

oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/auth/login")    #tokenUrl route de login utile pour Swagger

def get_current_user(token: str = Depends(oauth2_scheme)):
    try:
        payload = jwt.decode(
            token,
            settings.SECRET_KEY,
            algorithms=[settings.ALGORITHM]
        )
        user_id : str = payload.get("sub")
        
        if user_id is None:
            raise HTTPException(status_code=401, detail="Token invalide")

        return user_id

    except JWTError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid or expired token"
        )
