Backend/requirements.txt
fastapi
uvicorn[standard]

python-jose[cryptography]
passlib[bcrypt]

tinydb
python-dotenv

httpx
pydantic-settings
pydantic